-Refactoring files -Finish Home Screen ui
