+++ lib/features/home/presentation/home_screen.dart
@@ -1,13 +1,13 @@
 import 'package:doctor_app/core/helper/spacing.dart';
-import 'package:doctor_app/features/home/presentation/widgets/doctor_speciality_see_all.dart';
+import 'package:doctor_app/features/home/presentation/widgets/doctor_speciality/doctor_speciality_see_all.dart';
 import 'package:doctor_app/features/home/presentation/widgets/home_top_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import 'widgets/doctor_blue_container.dart';
-import 'widgets/doctor_speciality_list_view.dart';
-import 'widgets/doctors_recommendation_list_view.dart';
-import 'widgets/recommendation_doctor_and_see_all.dart';
+import 'widgets/banner/doctor_blue_container.dart';
+import 'widgets/doctor_speciality/doctor_speciality_list_view.dart';
+import 'widgets/recommendation_doctor/doctors_recommendation_list_view.dart';
+import 'widgets/recommendation_doctor/recommendation_doctor_and_see_all.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
