+++ lib/features/home/presentation/widgets/doctor_speciality/doctor_speciality_list_view.dart
@@ -1,0 +1,26 @@
+import 'package:doctor_app/core/helper/spacing.dart';
+import 'package:doctor_app/core/theming/colors.dart';
+import 'package:doctor_app/core/theming/styles.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:flutter_svg/svg.dart';
+
+import 'doctor_speciality_item.dart';
+
+class DoctorSpecialityListView extends StatelessWidget {
+  const DoctorSpecialityListView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 100.h,
+      child: ListView.builder(
+          scrollDirection: Axis.horizontal,
+          physics: BouncingScrollPhysics(),
+          itemCount: 8,
+          itemBuilder: (context, index) {
+            return DoctorSpecialityItem(index: index,);
+          }),
+    );
+  }
+}
