+++ lib/features/home/presentation/widgets/recommendation_doctor/doctors_recommendation_list_view.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'doctor_recommeded_card.dart';
+
+class DoctorsRecommendationListView extends StatelessWidget {
+  const DoctorsRecommendationListView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+      child: ListView.builder(
+          itemCount: 10,
+          itemBuilder: (context, index) {
+            return DoctorRecommededCard();
+          }),
+    );
+  }
+}
