+++ lib/features/home/presentation/widgets/recommendation_doctor/recommendation_doctor_and_see_all.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../core/theming/styles.dart';
+import '../../../../../core/theming/styles.dart';
 
 class RecommendationDoctorAndSeeAll extends StatelessWidget {
   const RecommendationDoctorAndSeeAll({super.key});
